add Result monad
