+++ ResultMonad/ResultMonad.csproj
@@ -1,0 +1,8 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+        <RootNamespace>ResultMonad</RootNamespace>
+    </PropertyGroup>
+
+</Project>
