+++ fp.sln
@@ -18,6 +18,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudImageSaver", "Tags
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization", "TagsCloudVisualization\TagsCloudVisualization.csproj", "{A701BA3A-4917-4F98-A917-7ECE16FCB42D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ResultMonad", "ResultMonad\ResultMonad.csproj", "{738110E6-664C-424F-938E-2459F55C861A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -52,11 +54,16 @@ Global
 		{A701BA3A-4917-4F98-A917-7ECE16FCB42D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A701BA3A-4917-4F98-A917-7ECE16FCB42D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A701BA3A-4917-4F98-A917-7ECE16FCB42D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{738110E6-664C-424F-938E-2459F55C861A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{738110E6-664C-424F-938E-2459F55C861A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{738110E6-664C-424F-938E-2459F55C861A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{738110E6-664C-424F-938E-2459F55C861A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{2E64584B-4FBC-4E7D-AEF2-003E8251307A} = {754C1CC8-A8B6-46C6-B35C-8A43B80111A0}
 		{0FA173DA-ED4F-4A4F-96B3-7EABAF795EB9} = {955A5D01-85E0-4549-BF22-1D1BF66CE1F6}
 		{D5668A88-AB6C-4447-9018-3F535F14AFCF} = {955A5D01-85E0-4549-BF22-1D1BF66CE1F6}
 		{A701BA3A-4917-4F98-A917-7ECE16FCB42D} = {955A5D01-85E0-4549-BF22-1D1BF66CE1F6}
+		{738110E6-664C-424F-938E-2459F55C861A} = {955A5D01-85E0-4549-BF22-1D1BF66CE1F6}
 	EndGlobalSection
 EndGlobal
